+++ react/typings/vtex.dreamstore-icons.d.ts
@@ -1,6 +1,6 @@
-declare module 'vtex.dreamstore-icons' {
-  import { ReactElement } from 'react'
+declare module "vtex.store-icons" {
+  import { ReactElement } from "react";
 
-  export const IconAssistantSales: ReactElement
-  export const IconProfile: ReactElement
+  export const IconAssistantSales: ReactElement;
+  export const IconProfile: ReactElement;
 }