+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - Add snapshot tests.
+- Using `store-icons` instead of `dreamstore-icons`
 
 ## [2.7.3] - 2019-02-14